Ensure float conversion for embeddings before numpy export
Updated get_text_embedding and get_image_embedding to explicitly convert embeddings to float before moving to CPU and exporting as numpy arrays. This ensures consistent data types and prevents potential issues with downstream processing.

+++ preprocessor/cli/commands/search.py
@@ -65,7 +65,7 @@ def get_text_embedding(text):
         embedding = output.hidden_states[-1][:, -1, :].squeeze(0)
         embedding = torch.nn.functional.normalize(embedding, p=2, dim=0)
 
-    return embedding.cpu().numpy().tolist()
+    return embedding.float().cpu().numpy().tolist()
 
 
 def get_image_embedding(image_path):
@@ -96,7 +96,7 @@ def get_image_embedding(image_path):
         embedding = output.hidden_states[-1][:, -1, :].squeeze(0)
         embedding = torch.nn.functional.normalize(embedding, p=2, dim=0)
 
-    return embedding.cpu().numpy().tolist()
+    return embedding.float().cpu().numpy().tolist()
 
 
 def load_hasher():
